pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scopescrape"
version = "0.1.0"
description = "Open-source community pain point discovery tool for Reddit and Hacker News"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Siva Pentakota", email = "sivasantosh143@gmail.com"},
]
keywords = ["reddit", "hacker-news", "pain-points", "community-intelligence", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "requests>=2.31,<3.0",
    "click>=8.1,<9.0",
    "pyyaml>=6.0,<7.0",
    "vaderSentiment>=3.3,<4.0",
    "rank-bm25>=0.2,<1.0",
    "python-dotenv>=1.0,<2.0",
    "rich>=13.0,<14.0",
]

[project.optional-dependencies]
nlp = [
    "spacy>=3.7,<4.0",
]
parquet = [
    "pyarrow>=15.0",
]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-cov>=5.0,<6.0",
    "ruff>=0.4,<1.0",
]
all = [
    "spacy>=3.7,<4.0",
    "pyarrow>=15.0",
]

[project.scripts]
scopescrape = "scopescrape.cli:main"

[project.urls]
Homepage = "https://scopescrape.earnedconviction.com"
Repository = "https://github.com/cvasan2sh/scopescrape_tool"
Documentation = "https://scopescrape.earnedconviction.com/docs"
Issues = "https://github.com/cvasan2sh/scopescrape_tool/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "live: tests that hit real APIs (deselect with -m 'not live')",
]
